# academe/server/openapi.yaml
openapi: 3.1.0
info:
  title: academe API
  version: "1"
  description: |
    Every error is {"error": {"code", "message", "requestId"}}. The app
    switches on `code`; `message` is for people.
    Access tokens last 15 minutes. When one is rejected with `invalid_token`,
    call /auth/refresh once; if that is rejected too, log the student out.
servers:
  - url: http://localhost:8080
paths:
...
  /study/decks:
    get:
      summary: Swipe decks for the student's class and board, with their progress. Empty until class and board are set.
      security: [{ bearer: [] }]
      parameters:
        - { name: subject, in: query, required: false, schema: { type: string }, description: "A subject id from /catalog/subjects" }
      responses:
        "200":
          description: Decks in subject, chapter and position order.
          content:
            application/json:
              schema:
                type: object
                required: [decks, chapters]
                properties:
                  decks: { type: array, items: { $ref: "#/components/schemas/DeckSummary" } }
                  chapters:
                    type: array
                    description: Every syllabus chapter for the student's class and board, including chapters with no lessons written yet, in subject and chapter order.
                    items: { $ref: "#/components/schemas/ChapterPlan" }
  /study/decks/{id}:
    get:
      summary: One lesson (deck) with all its cards, the student's kept cards and resume position. Every lesson opens on a start card and ends on a summary.
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The deck.
          content: { application/json: { schema: { $ref: "#/components/schemas/Deck" } } }
        "404":
          description: "`deck_not_found`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /study/decks/{id}/answers:
    post:
      summary: Grade an answer to a quiz card. A right answer pays QuizXP (5) once per card.
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [card, choice]
              properties:
                card: { type: integer, description: "Index of the quiz card in the deck" }
                choice: { type: integer, description: "Index of the chosen option" }
      responses:
        "200":
          description: Graded.
          content:
            application/json:
              schema:
                type: object
                required: [correct, answer, xpAwarded]
                properties:
                  correct: { type: boolean }
                  answer: { type: integer }
                  xpAwarded: { type: integer }
        "404":
          description: "`deck_not_found`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_card` or `invalid_choice`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /study/decks/{id}/completion:
    put:
      summary: Mark a deck finished with its quiz score. The best score is kept.
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [correct]
              properties:
                correct: { type: integer }
      responses:
        "204": { description: Saved. }
        "404":
          description: "`deck_not_found`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_score`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /study/decks/{id}/position:
    put:
      summary: Remember the card the student is on, for Resume.
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [card]
              properties:
                card: { type: integer }
      responses:
        "204": { description: Saved. }
        "404":
          description: "`deck_not_found`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_card`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /study/kept:
    put:
      summary: Keep a card for revision. Missed quiz questions are kept automatically when answered.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [deckId, card]
              properties:
                deckId: { type: string }
                card: { type: integer }
      responses:
        "204": { description: Kept. }
        "404":
          description: "`deck_not_found`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_card`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /study/kept/{deck}/{card}:
    delete:
      summary: Stop keeping a card.
      security: [{ bearer: [] }]
      parameters:
        - { name: deck, in: path, required: true, schema: { type: string } }
        - { name: card, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Removed. }
  /study/review:
    get:
      summary: Kept cards to revise. Without chapter, only the ones due now; with chapter, every kept card in it.
      security: [{ bearer: [] }]
      parameters:
        - { name: chapter, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Cards, soonest due first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/ReviewItem" } }
    post:
      summary: Rate a revised card. again = tomorrow; almost and knew push it further out.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [deckId, card, rating]
              properties:
                deckId: { type: string }
                card: { type: integer }
                rating: { type: string, enum: [again, almost, knew] }
      responses:
        "204": { description: Rescheduled. }
        "404":
          description: "`not_kept`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_rating`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /study/chapter-results:
    get:
      summary: The student's latest chapter test score per chapter.
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Results.
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results: { type: array, items: { $ref: "#/components/schemas/ChapterResult" } }
  /study/chapter-results/{id}:
    put:
      summary: Save a chapter test score. Each question's XP comes from /answers.
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [correct, total]
              properties:
                correct: { type: integer }
                total: { type: integer }
      responses:
        "204": { description: Saved. }
        "404":
          description: "`chapter_not_found`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422":
          description: "`invalid_score`"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: The access token from a Session. Any bearer route answers 401 `invalid_token` when the token has expired (15 minutes) or was issued before a password reset, a Google link or an account deletion request; refresh or log in again. Other server instances may accept a revoked token for up to 30 seconds.
...
  schemas:
...
    ReviewItem:
      type: object
      required: [deckId, card, reason, lessonTitle, chapterTitle, content]
      properties:
        deckId: { type: string }
        card: { type: integer }
        reason: { type: string, enum: [kept, missed] }
        lessonTitle: { type: string }
        chapterTitle: { type: string }
        content: { $ref: "#/components/schemas/DeckCard" }
    ChapterResult:
      type: object
      required: [chapterId, correct, total, completedAt]
      properties:
        chapterId: { type: string }
        correct: { type: integer }
        total: { type: integer }
        completedAt: { type: string, format: date-time }
...
    DeckSummary:
      type: object
      required: [id, chapterId, subject, subjectName, chapterNumber, chapterTitle, position, title, cards, quizzes, done, correct, resumeCard, kept]
      properties:
        id: { type: string }
        chapterId: { type: string }
        resumeCard: { type: integer, description: the card to resume on }
        kept: { type: integer, description: kept cards in this lesson }
        subject: { type: string }
        subjectName: { type: string }
        chapterNumber: { type: integer }
        chapterTitle: { type: string }
        position: { type: integer }
        title: { type: string }
        cards: { type: integer }
        quizzes: { type: integer }
        done: { type: boolean }
        correct: { type: integer }
    Deck:
      type: object
      required: [id, board, class, subject, chapterNumber, chapterTitle, position, title, language, cards]
      properties:
        id: { type: string }
        board: { type: string }
        class: { type: integer }
        subject: { type: string }
        chapterNumber: { type: integer }
        chapterTitle: { type: string }
        position: { type: integer }
        title: { type: string }
        language: { type: string }
        cards: { type: array, items: { $ref: "#/components/schemas/DeckCard" } }
        kept: { type: array, items: { type: integer }, description: "indexes of the cards this student kept" }
        resumeCard: { type: integer }
        status: { type: string, enum: [approved], description: "Generated lessons only; only approved lessons are served" }
        generatedBy:
          type: object
          description: Generated lessons only.
          properties:
            model: { type: string }
            promptVersion: { type: string }
            checkedAt: { type: string, format: date-time }
    ChapterPlan:
      type: object
      required: [id, subject, subjectName, number, title, unit, lessons]
      properties:
        id: { type: string, example: cbse-10-science-1 }
        subject: { type: string }
        subjectName: { type: string }
        number: { type: integer }
        title: { type: string }
        unit: { type: string, description: "Syllabus unit, empty when unknown" }
        formativeOnly: { type: boolean, description: "true for CBSE reading material that is assessed in school only and is not in the board exam; absent otherwise" }
        lessons:
          type: array
          description: Planned lessons in order. Unavailable ones are not written yet; show them as coming soon.
          items:
            type: object
            required: [id, position, title, available]
            properties:
              id: { type: string, example: cbse-10-science-1-1 }
              position: { type: integer }
              title: { type: string }
              available: { type: boolean, description: "true when GET /study/decks/{id} serves this lesson" }
    DeckCard:
      type: object
      required: [kind]
      properties:
        kind: { type: string, enum: [start, concept, table, example, quiz, summary] }
        goals: { type: array, items: { type: string }, description: start only }
        minutes: { type: integer, description: start only }
        rows: { type: array, items: { type: object, properties: { term: { type: string }, value: { type: string } } }, description: table only }
        steps: { type: array, items: { type: string }, description: example only, revealed one at a time }
        points: { type: array, items: { type: string }, description: summary only }
        remember: { type: string, description: concept or table, a highlighted line }
        title: { type: string, description: concept only }
        body: { type: string, description: "concept only. Lines starting \"- \" or \"1. \" are lists; **text** is bold" }
        question: { type: string, description: quiz only }
        options: { type: array, items: { type: string }, description: quiz only }
        answer: { type: integer, description: quiz only }
        why: { type: string, description: "quiz only; shown when the answer is wrong" }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object, description: "Only on some errors, e.g. limit_reached and pro_only: {feature, limit, resetsAt}." }